+++ pkg/resource/deploy/source_eval.go
@@ -819,19 +819,26 @@ func (rm *resmon) RegisterResource(ctx context.Context,
 		return nil, rpcerror.New(codes.Unavailable, "resource monitor shut down while waiting on step's done channel")
 	}
 
-	state := result.State
-	stable := result.Stable
-	var stables []string
-	for _, sta := range result.Stables {
-		stables = append(stables, string(sta))
+	// Filter out partially-known values if the requestor does not support them.
+	state, outputs := result.State, result.State.Outputs
+	if !req.GetSupportsPartialValues() {
+		logging.V(5).Infof("stripping unknowns from RegisterResource response for urn %v", state.URN)
+		filtered := resource.PropertyMap{}
+		for k, v := range outputs {
+			if !v.ContainsUnknowns() {
+				filtered[k] = v
+			}
+		}
+		outputs = filtered
 	}
+
 	logging.V(5).Infof(
-		"ResourceMonitor.RegisterResource operation finished: t=%v, urn=%v, stable=%v, #stables=%v #outs=%v",
-		state.Type, state.URN, stable, len(stables), len(state.Outputs))
+		"ResourceMonitor.RegisterResource operation finished: t=%v, urn=%v, #outs=%v",
+		state.Type, state.URN, len(outputs))
 
 	// Finally, unpack the response into properties that we can return to the language runtime.  This mostly includes
 	// an ID, URN, and defaults and output properties that will all be blitted back onto the runtime object.
-	obj, err := plugin.MarshalProperties(state.Outputs, plugin.MarshalOptions{
+	obj, err := plugin.MarshalProperties(outputs, plugin.MarshalOptions{
 		Label:        label,
 		KeepUnknowns: true,
 		KeepSecrets:  req.GetAcceptSecrets(),
@@ -840,11 +847,9 @@ func (rm *resmon) RegisterResource(ctx context.Context,
 		return nil, err
 	}
 	return &pulumirpc.RegisterResourceResponse{
-		Urn:     string(state.URN),
-		Id:      string(state.ID),
-		Object:  obj,
-		Stable:  stable,
-		Stables: stables,
+		Urn:    string(state.URN),
+		Id:     string(state.ID),
+		Object: obj,
 	}, nil
 }