+++ sdk/python/lib/pulumi/output.py
@@ -31,6 +31,7 @@ from typing import (
 
 from . import runtime
 from .runtime import known_types
+from .runtime import rpc
 
 if TYPE_CHECKING:
     from .resource import Resource
@@ -80,13 +81,18 @@ class Output(Generic[T]):
 
     def __init__(self, resources: Set['Resource'], future: Awaitable[T],
                  is_known: Awaitable[bool], is_secret: Optional[Awaitable[bool]] = None) -> None:
+        is_known = asyncio.ensure_future(is_known)
+        future = asyncio.ensure_future(future)
+
+        async def is_value_known() -> bool:
+            return await is_known and not contains_unknowns(await future)
 
         self._resources = resources
         self._future = future
-        self._is_known = is_known
+        self._is_known = asyncio.ensure_future(is_value_known())
 
         if is_secret is not None:
-            self._is_secret = is_secret
+            self._is_secret = asyncio.ensure_future(is_secret)
         else:
             self._is_secret = asyncio.Future()
             self._is_secret.set_result(False)
@@ -95,8 +101,13 @@ class Output(Generic[T]):
     def resources(self) -> Set['Resource']:
         return self._resources
 
-    def future(self) -> Awaitable[T]:
-        return self._future
+    def future(self, with_unknowns: Optional[bool] = None) -> Awaitable[T]:
+        # If the caller did not explicitly ask to see unknown values and the value of this output contains unnkowns,
+        # return None. This preserves compatibility with earlier versios of the Pulumi SDK.
+        async def get_value() -> T:
+            val = await self._future
+            return None if not with_unknowns and contains_unknowns(val) else val
+        return asyncio.ensure_future(get_value())
 
     def is_known(self) -> Awaitable[bool]:
         return self._is_known
@@ -105,7 +116,7 @@ class Output(Generic[T]):
         return self._is_secret
     # End private implementation details.
 
-    def apply(self, func: Callable[[T], Input[U]]) -> 'Output[U]':
+    def apply(self, func: Callable[[T], Input[U]], run_with_unknowns: Optional[bool] = None) -> 'Output[U]':
         """
         Transforms the data of the output with the provided func.  The result remains a
         Output so that dependent resources can be properly tracked.
@@ -124,39 +135,34 @@ class Output(Generic[T]):
         :return: A transformed Output obtained from running the transformation function on this Output's value.
         :rtype: Output[U]
         """
-        inner_is_known: asyncio.Future = asyncio.Future()
-        inner_is_secret: asyncio.Future = asyncio.Future()
-
-        # The "is_known" coroutine that we pass to the output we're about to create is derived from
-        # the conjunction of the two is_knowns that we know about: our own (self._is_known) and a future
-        # that we will resolve when running the apply.
-        async def is_known() -> bool:
-            inner = await inner_is_known
-            known = await self._is_known
-            return inner and known
-
-        # The "is_secret" coroutine that we pass to the output we're about to create is derived from
-        # the disjunction of the two is_secret that we know about: our own (self._is_secret) and a future
-        # that we will resolve when running the apply.
-        async def is_secret() -> bool:
-            inner = await inner_is_secret
-            secret = await self._is_secret
-            return inner or secret
+        result_is_known: asyncio.Future = asyncio.Future()
+        result_is_secret: asyncio.Future = asyncio.Future()
 
         # The "run" coroutine actually runs the apply.
         async def run() -> U:
             try:
+                # Await this output's details.
+                is_known = await self._is_known
+                is_secret = await self._is_secret
                 value = await self._future
+
                 if runtime.is_dry_run():
-                    # During previews only perform the apply if the engine was able to
-                    # give us an actual value for this Output.
-                    apply_during_preview = await self._is_known
+                    # During previews only perform the apply if the engine was able togive us an actual value for this
+                    # Output or if the caller is able to tolerate unknown values.
+                    apply_during_preview = is_known or run_with_unknowns
+
                     if not apply_during_preview:
                         # We didn't actually run the function, our new Output is definitely
                         # **not** known and **not** secret
-                        inner_is_known.set_result(False)
-                        inner_is_secret.set_result(False)
+                        result_is_known.set_result(False)
+                        result_is_secret.set_result(False)
                         return cast(U, None)
+
+                    # If we are running with unknown values and the value is explicitly unknown but does not actually
+                    # contain any unknown values, collapse its value to the unknown value. This ensures that callbacks
+                    # that expect to see unknowns during preview in outputs that are not known will always do so.
+                    if not is_known and run_with_unknowns and not contains_unknowns(value):
+                        value = UNKNOWN
 
                 transformed: Input[U] = func(value)
                 # Transformed is an Input, meaning there are three cases:
@@ -164,36 +170,34 @@ class Output(Generic[T]):
                 if isinstance(transformed, Output):
                     transformed_as_output = cast(Output[U], transformed)
                     # Forward along the inner output's _is_known and _is_secret values.
-                    inner_is_known.set_result(await transformed_as_output._is_known)
-                    inner_is_secret.set_result(await transformed_as_output._is_secret)
-                    return await transformed.future()
+                    result_is_known.set_result(await transformed_as_output._is_known)
+                    result_is_secret.set_result(await transformed_as_output._is_secret or is_secret)
+                    return await transformed.future(with_unknowns=True)
 
                 #  2. transformed is an Awaitable[U]
                 if isawaitable(transformed):
                     # Since transformed is not an Output, it is both known and not a secret.
-                    inner_is_known.set_result(True)
-                    inner_is_secret.set_result(False)
+                    result_is_known.set_result(True)
+                    result_is_secret.set_result(False)
                     return await cast(Awaitable[U], transformed)
 
                 #  3. transformed is U. It is trivially known.
-                inner_is_known.set_result(True)
-                inner_is_secret.set_result(False)
+                result_is_known.set_result(True)
+                result_is_secret.set_result(False)
                 return cast(U, transformed)
             finally:
                 # Always resolve the future if it hasn't been done already.
-                if not inner_is_known.done():
+                if not result_is_known.done():
                     # Try and set the result. This might fail if we're shutting down,
                     # so swallow that error if that occurs.
                     try:
-                        inner_is_known.set_result(False)
-                        inner_is_secret.set_result(False)
+                        result_is_known.set_result(False)
+                        result_is_secret.set_result(False)
                     except RuntimeError:
                         pass
 
         run_fut = asyncio.ensure_future(run())
-        is_known_fut = asyncio.ensure_future(is_known())
-        is_secret_fut = asyncio.ensure_future(is_secret())
-        return Output(self._resources, run_fut, is_known_fut, is_secret_fut)
+        return Output(self._resources, run_fut, result_is_known, result_is_secret)
 
     def __getattr__(self, item: str) -> 'Output[Any]':
         """
@@ -203,7 +207,7 @@ class Output(Generic[T]):
         :return: An Output of this Output's underlying value's property with the given name.
         :rtype: Output[Any]
         """
-        return self.apply(lambda v: getattr(v, item))
+        return self.apply(lambda v: UNKNOWN if isinstance(v, Unknown) else getattr(v, item), True)
 
 
     def __getitem__(self, key: Any) -> 'Output[Any]':
@@ -214,7 +218,7 @@ class Output(Generic[T]):
         :return: An Output of this Output's underlying value, keyed with the given key as if it were a dictionary.
         :rtype: Output[Any]
         """
-        return self.apply(lambda v: v[key])
+        return self.apply(lambda v: UNKNOWN if isinstance(v, Unknown) else v[key], True)
 
     @staticmethod
     def from_input(val: Input[T]) -> 'Output[T]':
@@ -226,16 +230,17 @@ class Output(Generic[T]):
         :return: A deeply-unwrapped Output that is guaranteed to not contain any Input values.
         :rtype: Output[T]
         """
+
         # Is it an output already? Recurse into the value contained within it.
         if isinstance(val, Output):
-            return val.apply(Output.from_input)
+            return val.apply(Output.from_input, True)
 
         # Is a dict or list? Recurse into the values within them.
         if isinstance(val, dict):
             # Since Output.all works on lists early, serialize this dictionary into a list of lists first.
             # Once we have a output of the list of properties, we can use an apply to re-hydrate it back into a dict.
             transformed_items = [[k, Output.from_input(v)] for k, v in val.items()]
-            return Output.all(*transformed_items).apply(lambda props: {k: v for k, v in props})
+            return Output.all(*transformed_items).apply(lambda props: {k: v for k, v in props}, True)
 
         if isinstance(val, list):
             transformed_items = [Output.from_input(v) for v in val]
@@ -251,7 +256,7 @@ class Output(Generic[T]):
         # as the value future for a new output.
         if isawaitable(val):
             promise_output = Output(set(), asyncio.ensure_future(val), is_known_fut, is_secret_fut)
-            return promise_output.apply(Output.from_input)
+            return promise_output.apply(Output.from_input, True)
 
         # Is it a prompt value? Set up a new resolved future and use that as the value future.
         value_fut = asyncio.Future()
@@ -307,7 +312,7 @@ class Output(Generic[T]):
 
         # gather_futures, which aggregates the list of futures in each input to a future of a list.
         async def gather_futures(outputs):
-            value_futures = list(map(lambda o: asyncio.ensure_future(o.future()), outputs))
+            value_futures = list(map(lambda o: asyncio.ensure_future(o.future(with_unknowns=True)), outputs))
             return await asyncio.gather(*value_futures)
 
         # First, map all inputs to outputs using `from_input`.
@@ -341,3 +346,21 @@ class Output(Generic[T]):
 
         transformed_items = [Output.from_input(v) for v in args]
         return Output.all(*transformed_items).apply("".join)
+
+
+@known_types.unknown
+class Unknown:
+    """
+    Unknown represents a value that is unknown.
+    """
+
+    def __init__(self):
+        pass
+
+UNKNOWN = Unknown()
+"""
+UNKNOWN is the singleton unknown value.
+"""
+
+def contains_unknowns(val: Any) -> bool:
+    return rpc.contains_unknowns(val)