+++ CHANGELOG.md
@@ -8,6 +8,12 @@ CHANGELOG
   used.
   [#3459](https://github.com/pulumi/pulumi/pull/3459)
 
+- Re-apply "propagate resource inputs to resource state during preview, including first-class unknown values." The new
+  set of changes have additional fixes to ensure backwards compatibility with earlier code. This allows the preview to
+  better estimate the state of a resource after an update, including property values that were populated using defaults
+  calculated by the provider.
+  [#3327](https://github.com/pulumi/pulumi/pull/3327)
+
 ## 1.5.1 (2019-11-06)
 
 - Include the .NET language provider in the Windows SDK.